Refactored folder structure to organize communication modules under dedicated service directories.
