+++ Hardware/ESPIDF-Server/main/mqtt_service/mqtt_communication.c
